+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - DefaultToggleButtonBuilder, AbstractToggleMenuItemBuilder bug fixed, ObservableStateLink added, now sets initial value.
 - Control.enabled() now returns Optional, AbstractControl, unnecessary enabledObserver no longer created, unnecessary overload of getValue() removed.
 - AbstractDialogBuilder bug fixed, now handles null frame icon.
+- FilterTableCellEditor.Builder.resizeRow() added for overriding the default table setting.
 ### is.codion.swing.framework.ui
 - EntityTableExportModel.all() added, EntityTableExportPanel all/selected radio button setup improved.